machine_learning_scenarios/requirements.txt
sagemaker
python-decouple
pandas